add comments for each IOS/IOS-XE regex entry
Debugging the IOS/IOS-XE filters is much easier if you know what was
supposed to be matched by each regex. You can easily see if the code
was even supposed to match what you want matched... or debug the line
that is supposed to be matching. This commit just adds comments, no
"code" changes.